refactor evaluator, set def value for SHOW_RL_OBS
