+++ projects/habitat_ovmm/README.md
@@ -44,7 +44,7 @@ unzip hab_stretch_v1.0.zip
 
 Run
 ```
-python projects/habitat_ovmm/eval_dataset.py
+python projects/habitat_ovmm/eval_baselines_agent.py
 ```
 
 Results are saved to `datadump/images/eval_hssd/`.
@@ -103,27 +103,31 @@ python -u -m habitat_baselines.run \
 
 ## Running evaluations
 
+
+### With GT semantics
 ```
-cd /path/to/home-robot
+python projects/habitat_ovmm/eval_baselines_agent.py
+```
 
-
-# Evaluation on complete episode dataset with GT semantics
-python projects/habitat_ovmm/eval_dataset.py
-
-# Evaluation on complete episode dataset with DETIC
+### With DETIC
 Ensure `GROUND_TRUTH_SEMANTICS:0` in `configs/agent/hssd_eval.yaml` before running the above command
 
-# Evaluation on specific episodes
-python projects/habitat_ovmm/eval_dataset.py habitat.dataset.episode_ids="[151,182]"
-
-# Evaluating all baseline variants
-# 1. First generate all possible configs using the base config `configs/agent/hssd_eval.yaml`. Configs will be saved under `projects/habitat_ovmm/configs/agent/generated`
-python projects/habitat_ovmm/scripts/gen_configs.py
-
-# 2. Run evaluation using the generated config files
-python projects/habitat_ovmm/eval_dataset.py --baseline_config_path projects/habitat_ovmm/configs/agent/generated/<dir_name>/<manip>_m_<nav>_n_<perception><viz?>.yaml
-Here <manip>/<nav> are to be set to 'h' or 'r' for heuristic and RL skills respectively. <perception> is one of 'gt'/'detic'. Append <viz?>='_viz' for saving images.
-
+### On specific episodes
 ```
+python projects/habitat_ovmm/eval_baselines_agent.py habitat.dataset.episode_ids="[151,182]"
+```
+
+### Evaluating all baseline variants
+1. First generate all possible configs using the base config `configs/agent/hssd_eval.yaml`. Configs will be saved under `projects/habitat_ovmm/configs/agent/generated`
+```
+python projects/habitat_ovmm/scripts/gen_configs.py
+```
+
+2. Run evaluation using the generated config files
+```
+python projects/habitat_ovmm/eval_baselines_agent.py --baseline_config_path projects/habitat_ovmm/configs/agent/generated/<dir_name>/<manip>_m_<nav>_n_<perception><viz?>.yaml
+```
+
+Here <manip>/<nav> are to be set to 'h' or 'r' for heuristic and RL skills respectively. <perception> is one of 'gt'/'detic'. Append <viz?>='_viz' for saving images.
 
 In case you run into issues, please prepend your python command with `HABITAT_ENV_DEBUG=1` to get a better error message.
